refactor: replace bundled ayumi modules with npm package imports
Add `ayumi` as a git dependency (`github:yama-kei/ayumi#main`) and
remove the 8 pipeline/vault/fetcher modules that are now provided by
the ayumi package. The remaining 4 integration-glue modules import
`Topic` from ayumi and inline the MPG-specific pending-review and
Drive folder management logic that has no ayumi equivalent.

Files removed:
- extraction-pipeline.ts, vault-writer.ts, drive-writer.ts
- web-fetcher.ts, local-reader.ts, topic-summarizer.ts
- life-context-setup.ts, exclusions.ts

Files kept (glue):
- index.ts — re-exports ayumi primitives + MPG glue
- curator-commands.ts — inlined vault/drive pending review logic
- life-context-loader.ts — imports Topic from ayumi
- presets.ts — unchanged
- config/ — unchanged

Closes #202
Ref: yama-kei/ayumi#44
