+++ package.json
@@ -17,6 +17,7 @@
     "test:watch": "vitest"
   },
   "dependencies": {
+    "ayumi": "github:yama-kei/ayumi#main",
     "discord.js": "^14.14.0",
     "dotenv": "^16.4.0"
   },
