+++ src/ayumi/index.ts
@@ -3,12 +3,19 @@
  *
  * MPG core should only import from this file. Everything else
  * in src/ayumi/ is an internal implementation detail.
+ *
+ * Pipeline primitives (fetchUrl, extractContent, summarizeContent,
+ * createVaultWriter, createDriveBroker, etc.) are available via
+ * `import { ... } from 'ayumi'` directly.
  */
 
 import { loadLifeContext } from './life-context-loader.js';
 
 export { loadLifeContext };
-export { writeTopicToVault, type VaultWriterOptions, type VaultWriteResult } from './vault-writer.js';
+
+// Re-export ayumi pipeline primitives that other MPG modules may need
+export type { Topic, DriveBroker, ExtractedContent, ClassificationResult, ArticleSummary, LLMComplete } from 'ayumi';
+export { fetchUrl, extractContent, classifyContent, summarizeContent, createVaultWriter, createDriveBroker, parseFile } from 'ayumi';
 
 /**
  * Get agent context for the given agent name.
